Revert "[InstSimplify] Bypass no-op `and`-mask, using known bits (PR49543)"
This reverts commit 9fcc45350e272bb54efe33ede88247be07cfbabb.

While this is strictly more powerful, it is also strictly slower.
InstSimplify intentionally does not perform many folds that it
is allowed to perform, if doing so requires a KnownBits calculation
that will be repeated in InstCombine.

Maybe it's worthwhile to do this here, but that needs a more
explicitly stated motivation, evaluated in a review.