Reapply "[ORC] Add unit tests for parts of the ..." with fixes and improvements.
This reapplies 2d15bdda649, which was reverted in ec64d797599 due to buildbot
errors.

This version checks that a JIT instance can be safely constructed, skipping
tests if it can not be. To enable this it introduces new C API to retrieve and
set the target triple for a JITTargetMachineBuilder.