Reland [X86] X86TTIImpl::getInterleavedMemoryOpCostAVX2(): use getMemoryOpCost()
Now that getMemoryOpCost() correctly handles all the vector variants,
we should no longer hand-roll our own version of it, but use it directly.

The AVX512 variant probably needs a similar change,
but there it is less obvious.

This was initially landed in 6700561dd52f9a4377215bac09043d52b84ba6d6,
but was reverted in 66317a26514375ac76c5766b0ea3d89035fb0222
because the patch it depends on was reverted.