[InstCombine] visitUnreachableInst(): iteratively erase instructions leading to unreachable
In the original review D87149 it was mentioned that this approach was tried,
and it lead to infinite combine loops, but i'm not seeing anything like that now,
neither in the `check-llvm`, nor on some codebases i tried.

This is a recommit of 8dfabdbf73371e916390aa51edbe788d953cd336,
which i immediately reverted because i have messed up something
during branch switch, and 6fc1cde8d0b1e2f6176457cca3545b1da8a8370e
accidentally ended up being pushed, which was very much not the intention.

Reviewed By: spatel

Differential Revision: https://reviews.llvm.org/D105339