llvm-symbolizer: Fix "start file" to work with Split DWARF
A followup to the feature added in
a696c9169313ba8aa6abf95794a624f1d8984a5c that added the optional "start
file name" to match "start line" - but this didn't work with Split DWARF
because of the need for the decl file number resolution code to refer
back to the skeleton unit to find its .debug_line contribution. So this
patch adds the necessary infrastructure to track the skeleton unit
corresponding to a split full unit for the purpose of this lookup.