[InstCombine] reorder icmp with offset folds for better results
This set of folds was added recently with:
aef5c74133cd
2ec6b0ace86c
b01a60448a96

...and I noted that this wasn't likely to fire in code derived
from C/C++ source because of nsw in particular. But I didn't
notice that I had placed the code above the no-wrap block
of transforms.

This is likely the cause of regressions noted from the previous
commit because -- as shown in the test diffs -- we may have
transformed into a compare with an arbitrary constant rather
than a simpler signbit test.