Recommit "[VPlan] Add recipe for first-order rec phis, make splicing explicit."
This reverts the revert commit 3dc38429f0dd92ad496117e1e294379d8bb7db6b.

The patch originally got reverted due to a crash:
https://bugs.chromium.org/p/chromium/issues/detail?id=1232798#c2

The underlying issue was that we were not using the stored values from
the modified memory recipes, but the out-of-date values directly from
the IR (accessed via the VPlan). This should be fixed in 11dc2caad. A
reduced version of the reproducer has been added in 19b149162b61.